feat: resolve PR comments
